Final system tweaks, performance optimisation
Changelog:
 - Asteroids: removed gravimetrics component to reduce performance hit
 - Sol: added Luna and new Jupiter
 - Spawner: now randomises scale and rotation, and applies a nomenclatured name
 - MenuToggle: commented and added key templating
 - Stations: removed relativistics component to improve performance
 - Planets: removed relativistics component to improve performance

+++ Assets/Scripts/Utility/QualityControl.cs
@@ -62,7 +62,7 @@ public class QualityControl : MonoBehaviour
             mCamera.GetComponent<FlareLayer>().enabled = true;
         }
 
-        QualitySettings.SetQualityLevel((int)newLevel);
+        QualitySettings.SetQualityLevel(newLevel);
         quality = (QualityLevel)newLevel;
 
         if (vSyncToggle)
